:sparkles: [feat][be] 매크로 트랜잭션 추가 및 세부 버그 수정 - 각 api들의 엔드포인트를 /api로 시작하게끔 통일하였습니다. - 매크로 엔드포인트 반환값을 dto로 수정하였습니다. - 매크로 실행 부분에 트랜잭션이 동작하도록 추가하였습니다. (비동기동작) - get executionId를 통해 처리 상황을 확인할 수 있습니다 (폴링 방식으로 구현) - 스웨거를 통해 각 api의 테스트를 완료하였습니다. - 세부 변수이름들을 컨벤션에 맞게 변경하였습니다. - issue: #2, #10

+++ hyo-banking-be/src/main/java/org/atmgigi/hyobankingbe/account/dto/AccountInfoResponseDTO.java
@@ -2,6 +2,8 @@ package org.atmgigi.hyobankingbe.account.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record AccountInfoResponseDTO(
         long userId,
@@ -10,6 +12,6 @@ public record AccountInfoResponseDTO(
         String accountNo,
         String accountType,
         String currencyCode,
-        int balanceCache
+        BigDecimal balanceCache
 ) {
 }
